Issued test skeleton created
